refactor(multitable): extract permission service
